added perf_mon,uptime keys and fixed some functions for win32

+++ include/perfmon.h
@@ -43,16 +43,17 @@
 struct perfcounter
 {
 	struct perfcounter *next;
-	DWORD pdhIndex;
-	char name[MAX_PERFCOUNTER_NAME_LEN];	/* must be caracter array! if you want to rewrite */
-						/* to dinacal memory allocation CHECK for usage   */
-						/* of sizeof function                             */
+	unsigned long	pdhIndex;
+	char		name[MAX_PERFCOUNTER_NAME_LEN];	
+	/* must be caracter array! if you want to rewrite */
+	/* to dinacal memory allocation CHECK for usage   */
+	/* of sizeof function                             */
 };
 
 typedef struct perfcounter PERFCOUNTER;
 
 extern PERFCOUNTER *PerfCounterList;
 
-char *GetCounterName(DWORD index);
+char *GetCounterName(unsigned long index);
 
 #endif /* ZABBIX_PERFMON_H */
